pyproject.toml
[tool.poetry]
name = "reddit-daily-news"
version = "0.1.0"
description = ""
authors = ["Xunjie Liu <xunjieliu97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
praw = "^7.8.1"
loguru = "^0.7.3"
pytest = "^8.3.5"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pygithub = "^2.6.1"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
